feat: add create_stop_sign and create_speed_limit functions to gui

+++ src/constants.rs
@@ -1,6 +1,6 @@
-pub static MPS_TO_KPH: f32 = 3.6;
-pub static MPS_TO_MPH: f32 = 2.237;
+pub static MPS_TO_KPH: f64 = 3.6;
+pub static MPS_TO_MPH: f64 = 2.237;
 pub static CHAR_MAP_SIZE: usize = 40;
-pub static WORLD_SIZE: f32 = 200.0;
-pub static METERS_TO_KM: f32 = 0.001;
-pub static METERS_TO_MILES: f32 = 0.000621371;
+pub static WORLD_SIZE: f64 = 200.0;
+pub static METERS_TO_KM: f64 = 0.001;
+pub static METERS_TO_MILES: f64 = 0.000621371;
